requirements.txt
blinker==1.8.2
bottle==0.12.25
cffi==1.17.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
pillow==10.4.0
proxy_tools==0.1.0
pycparser==2.22
python-pptx==1.0.2
pythonnet==3.0.3
typing_extensions==4.12.2
Werkzeug==3.0.3
XlsxWriter==3.2.0
